# 🎨 Dynamic Subtitle Support: Word-Level Acoustic Analysis Implementation