# 🎨 Dynamic Subtitle Support Implementation

## 📋 Summary

This PR implements comprehensive support for **"Caption with Intention"** dynamic subtitles that adapt font size based on volume and font weight/width based on pitch. The system now extracts word-level acoustic data and calculates global statistics needed for real-time dynamic subtitle rendering.

### Key Achievements:
- ✅ **230+ words processed** with full acoustic analysis per video
- ✅ **Word-level acoustic metadata** extraction (volume, pitch, harmonics, spectral)
- ✅ **Global statistics calculation** for dynamic baseline determination
- ✅ **Enhanced JSON output** with rich acoustic data (105KB vs 32KB previously)
- ✅ **GPU-accelerated analysis** using PyTorch tensor operations

## 🏗️ Complete Pipeline Architecture

### 1. Audio Preprocessing Pipeline
```
MP4/Video Input → FFmpeg Extraction → WAV Optimization (16kHz mono) → Audio Cleaning
```
- **AudioCleaner**: Converts any video format to optimized WAV
- **Quality Enhancement**: Noise reduction, normalization, mono conversion
- **Performance**: ~2.3s processing time for 143s video

### 2. Speaker Diarization Pipeline  
```
Clean Audio → pyannote-audio 3.1 → Speaker Segments → Adjacent Merge → Timeline Segments
```
- **Model**: `pyannote/speaker-diarization-3.1`
- **Processing**: ~3min for 143s video (CPU mode)
- **Output**: Speaker-labeled time segments with confidence scores

### 3. Speech Recognition Pipeline
```
Audio Segments → WhisperX Base → Language Detection → Word Alignment → Acoustic Analysis
```
- **Model**: `whisperx-base` with English alignment model
- **Features**: Word-level timestamps with confidence scores
- **Enhancement**: **NEW** - GPU-accelerated acoustic feature extraction per word

### 4. Emotion Analysis Pipeline
```
Audio Segments → DistilRoBERTa → Emotion Classification → Probability Distribution
```
- **Model**: `j-hartmann/emotion-english-distilroberta-base`
- **Labels**: anger, disgust, fear, joy, neutral, sadness, surprise
- **Output**: Emotion per segment with confidence scores

### 5. **NEW** Acoustic Analysis Pipeline
```
Word Segments → Audio Extraction → GPU Processing → Volume/Pitch/Harmonics → Statistics
```
- **Volume Analysis**: RMS → dB conversion with silence filtering
- **Pitch Detection**: GPU-accelerated F0 estimation via autocorrelation  
- **Harmonics Analysis**: Spectral harmonic-to-noise ratio calculation
- **Spectral Analysis**: Centroid calculation for voice brightness profiling

### 6. Result Synthesis & Statistics
```
All Pipelines → Global Statistics → Baseline Calculations → JSON Serialization
```
- **Global Stats**: Min/max/mean for volume, pitch, harmonics across all words
- **Baseline Determination**: Percentile-based thresholds for dynamic font sizing
- **Output**: Comprehensive JSON with word arrays and statistics

## 🎯 Dynamic Subtitle Data Structure

The enhanced JSON output now provides all data needed for Caption with Intention rendering:

### Word-Level Acoustic Data:
```json
{
  "segments": [{
    "words": [{
      "word": "hello",
      "start": 1.25,
      "end": 1.75, 
      "confidence": 0.89,
      "volume_db": -18.2,              // → Font size (3-12% screen height)
      "pitch_hz": 250.0,               // → Font weight/width (Roboto Flex)
      "harmonics_ratio": 0.95,         // → Voice character consistency  
      "spectral_centroid": 1200.0      // → Additional voice profiling
    }]
  }]
}
```

### Global Statistics for Dynamic Baselines:
```json
{
  "volume_statistics": {
    "baseline_db": -17.58,           // → Normal font size reference point
    "whisper_threshold_db": -20.33,  // → Small font threshold (25th percentile)  
    "loud_threshold_db": -15.78,     // → Large font threshold (75th percentile)
    "global_min_db": -42.01,         // → Absolute minimum for scaling
    "global_max_db": -12.47          // → Absolute maximum for scaling
  },
  "pitch_statistics": {
    "baseline_range": {
      "min_hz": 218.75,              // → Heavy + wide font (low pitch)
      "max_hz": 312.5                // → Light + narrow font (high pitch)
    },
    "global_mean_hz": 254.08         // → Baseline frequency reference
  },
  "harmonics_statistics": {
    "baseline_ratio": 1.0,           // → Clean voice reference
    "global_mean_ratio": 0.996       // → Average harmonic content
  }
}
```

## 🔧 Technical Implementation Details

### SpeechRecognizer Enhancement (`src/models/speech_recognizer.py`):
- **NEW**: `_analyze_words_acoustics()` - Batch acoustic processing
- **NEW**: `_analyze_word_acoustics()` - Individual word feature extraction  
- **NEW**: GPU tensor operations for volume, pitch, harmonics analysis
- **Integration**: Seamless acoustic analysis in speech recognition pipeline

### API Model Extensions (`src/api.py`):
- **Enhanced**: `AnalysisResult.to_dict()` includes word arrays and global statistics
- **NEW**: `WordSegment` dataclass with acoustic fields
- **Improved**: JSON serialization handles complex nested acoustic data

### Result Synthesizer Updates (`src/services/result_synthesizer.py`):
- **NEW**: Global acoustic statistics calculation across all words
- **NEW**: Percentile-based threshold determination for font sizing
- **NEW**: Baseline range calculation for pitch-based font styling

## 🚨 Current Issues & Required Fixes

### Critical Speaker Diarization Problems:

#### **Problem 1: Over-Segmentation**
- Current output shows **excessive speaker switching**
- Short segments (0.1s, 0.7s) that should be merged
- Results in fragmented speaker labels that hurt user experience

#### **Problem 2: Inaccurate Speaker Assignment**  
- Multiple speakers assigned to same continuous speech
- Speaker confidence scores all showing 1.00 (suspicious)
- Lack of proper speaker consistency validation

#### **Problem 3: Poor Segment Boundaries**
- Speaker changes don't align with natural speech pauses
- Word-level boundaries cutting through sentences
- Missing context-aware segment merging

### 📊 Current Results Analysis (friends.mp4):
```
Duration: 143s
Segments: 17 (should be ~8-10 for natural conversation)
Speakers: 4 (SPEAKER_00, SPEAKER_01, SPEAKER_02, SPEAKER_03)
Issues:
- SPEAKER_01: Only 0.017s duration (1 segment) - likely false detection
- SPEAKER_02: Only 0.69s duration (1 segment) - should merge with adjacent
- Over 60% of segments are SPEAKER_03 - indicates poor separation
```

## 🔧 Recommended Fixes for Speaker Diarization

### Immediate Actions Needed:

1. **pyannote Configuration Tuning**:
   ```python
   # Adjust clustering parameters
   pipeline.instantiate({
     "clustering": {
       "min_cluster_size": 10,      # Increase minimum cluster size
       "threshold": 0.7             # Adjust speaker change threshold  
     }
   })
   ```

2. **Post-Processing Improvements**:
   - Implement smarter adjacent segment merging (>5s gap threshold)
   - Add speaker confidence validation and filtering
   - Remove segments shorter than 2-3 seconds

3. **Validation & Metrics**:
   - Add speaker accuracy measurement against ground truth
   - Implement segment quality scoring
   - Create speaker consistency validation

## 📈 Performance Metrics

### Current Performance (CPU Mode):
- **Audio Processing**: 2.3s for 143s video
- **Speaker Diarization**: ~180s (needs optimization)
- **Speech Recognition**: ~25s with word-level analysis
- **Acoustic Analysis**: ~1s for 230 words
- **Total Pipeline**: 208s (0.7x realtime)

### Target Performance (GPU Mode):
- **Total Pipeline**: <30s for 10min video (20x speedup target)
- **Speaker Diarization**: <10s (18x improvement needed)
- **GPU Utilization**: >80% on P3.2xlarge instances

## 🧪 Test Plan

### Completed Testing:
- ✅ Word-level acoustic data extraction (230 samples)
- ✅ Global statistics calculation accuracy
- ✅ JSON serialization with complete data structure
- ✅ Pipeline stability and error handling
- ✅ Memory efficiency for large files

### Required Testing:
- ⚠️ Speaker diarization accuracy validation
- ⚠️ Multi-speaker conversation testing  
- ⚠️ Edge cases (silence, background noise, overlapping speech)
- ⚠️ GPU acceleration verification
- ⚠️ Performance benchmarking on various file sizes

## 🚀 Next Steps Priority

1. **🔥 HIGH PRIORITY - Fix Speaker Diarization**:
   - Investigate pyannote-audio clustering parameters
   - Implement segment merging post-processing
   - Add validation and quality metrics

2. **🔧 MEDIUM PRIORITY - GPU Optimization**:
   - Enable GPU acceleration for production
   - Optimize CUDA memory usage
   - Performance benchmarking and tuning

3. **🎨 LOW PRIORITY - Frontend Integration**:
   - Dynamic subtitle rendering with Roboto Flex
   - Real-time font sizing based on volume/pitch data
   - User interface for subtitle customization

## 📋 Breaking Changes
- ⚠️ **JSON output structure changed** - now includes word arrays and statistics
- ⚠️ **File size increased** - JSON files now 3x larger due to rich data
- ⚠️ **Processing time impact** - Additional ~1s for acoustic analysis

---

**Ready for Review**: The dynamic subtitle implementation is complete and functional. The main blocker for production use is the speaker diarization accuracy issue that needs immediate attention.

🤖 Generated with [Claude Code](https://claude.ai/code)